refactor(imaging-api): rename test files to test_{source_file}.py convention
- test_imaging_endpoints.py → test_imaging.py
- test_download_endpoints.py → test_download.py
- test_upload_endpoints.py → test_upload.py
- test_users_endpoints.py → test_users.py
- test_retrieval_endpoints.py → test_retrieval.py
- Merge test_projects_endpoints.py into test_projects.py
- Merge test_upload_extended.py into test_upload.py
